requirements.txt
# 核心依赖
torchvision
opencv-python
numpy
ffmpeg-python
einops
formulaic
funcsigs
future
gitdb
gitpython
interface-meta
iopath
jpeg4py
jsonpatch
jsonpointer
latex
libarchive-c
linearmodels
lmdb
loguru
mat73
memory-profiler
msgpack
ninja
pathtools
promise
property-cached
protobuf
pycocotools
pyhdfe
ruamel.yaml
sentry-sdk
setproctitle
setuptools-scm
shapely
shortuuid
shutilwhich
smmap
tables
tempdir
tensorboardx
thop
tikzplotlib
timm
tomli
torchfile
visdom
wandb
webcolors
yaspin